added regex to skip a incomplete FETCH or TAG line
sometimes a FETCH or TAG line is parsed while it isn't a FETCH or TAG
line. This little regex tries to prevent that.

For example: ` * branch feature/PROJECT-12` this leads to an Exception
that the $flags variable is undefined. Which is true because there is no
flag in that line.

Signed-off-by: Tonko Mulder <tonko@tonkomulder.nl>